fix: revert BigInt change - Square SDK types require bigint

+++ src/app/api/stripe/create-payment-intent/route.ts
@@ -165,7 +165,7 @@ async function trySquarePayment(
       sourceId,
       idempotencyKey,
       amountMoney: {
-        amount: Math.round(amount * 100),
+        amount: BigInt(Math.round(amount * 100)),
         currency: currency.toUpperCase() as any,
       },
       locationId,
